+++ tests/Application/AbstractGetJobTest.php
@@ -4,9 +4,6 @@ declare(strict_types=1);
 
 namespace App\Tests\Application;
 
-use App\Entity\Job;
-use App\Repository\JobRepository;
-use App\Services\UlidFactory;
 use SmartAssert\TestAuthenticationProviderBundle\ApiTokenProvider;
 use Symfony\Component\Uid\Ulid;
 
@@ -76,44 +73,5 @@ abstract class AbstractGetJobTest extends AbstractApplicationTest
                 'apiToken' => 'invalid api token',
             ],
         ];
-    }
-
-    public function testGetSuccess(): void
-    {
-        $apiTokenProvider = self::getContainer()->get(ApiTokenProvider::class);
-        \assert($apiTokenProvider instanceof ApiTokenProvider);
-        $apiToken = $apiTokenProvider->get('user@example.com');
-
-        $ulidFactory = self::getContainer()->get(UlidFactory::class);
-        \assert($ulidFactory instanceof UlidFactory);
-
-        $suiteId = $ulidFactory->create();
-
-        $jobRepository = self::getContainer()->get(JobRepository::class);
-        \assert($jobRepository instanceof JobRepository);
-        self::assertCount(0, $jobRepository->findAll());
-
-        $createResponse = self::$staticApplicationClient->makeCreateJobRequest($apiToken, $suiteId, 600);
-
-        self::assertSame(200, $createResponse->getStatusCode());
-        self::assertSame('application/json', $createResponse->getHeaderLine('content-type'));
-
-        $createResponseData = json_decode($createResponse->getBody()->getContents(), true);
-        self::assertIsArray($createResponseData);
-        self::assertArrayHasKey('id', $createResponseData);
-        self::assertTrue(Ulid::isValid($createResponseData['id']));
-        $jobId = $createResponseData['id'];
-
-        $getResponse = self::$staticApplicationClient->makeGetJobRequest($apiToken, $jobId);
-
-        self::assertSame(200, $getResponse->getStatusCode());
-        self::assertSame('application/json', $getResponse->getHeaderLine('content-type'));
-
-        $job = $jobRepository->find($jobId);
-        self::assertInstanceOf(Job::class, $job);
-
-        $responseData = json_decode($getResponse->getBody()->getContents(), true);
-
-        self::assertEquals($job->jsonSerialize(), $responseData);
     }
 }
