+++ tests/Functional/Application/GetJobSuccessTest.php
@@ -1,0 +1,56 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Functional\Application;
+
+use App\Entity\Job;
+use App\Repository\JobRepository;
+use App\Services\UlidFactory;
+use App\Tests\Application\AbstractApplicationTest;
+use SmartAssert\TestAuthenticationProviderBundle\ApiTokenProvider;
+use Symfony\Component\Uid\Ulid;
+
+class GetJobSuccessTest extends AbstractApplicationTest
+{
+    use GetClientAdapterTrait;
+
+    public function testGetSuccess(): void
+    {
+        $apiTokenProvider = self::getContainer()->get(ApiTokenProvider::class);
+        \assert($apiTokenProvider instanceof ApiTokenProvider);
+        $apiToken = $apiTokenProvider->get('user@example.com');
+
+        $ulidFactory = self::getContainer()->get(UlidFactory::class);
+        \assert($ulidFactory instanceof UlidFactory);
+
+        $suiteId = $ulidFactory->create();
+
+        $jobRepository = self::getContainer()->get(JobRepository::class);
+        \assert($jobRepository instanceof JobRepository);
+        self::assertCount(0, $jobRepository->findAll());
+
+        $createResponse = self::$staticApplicationClient->makeCreateJobRequest($apiToken, $suiteId, 600);
+
+        self::assertSame(200, $createResponse->getStatusCode());
+        self::assertSame('application/json', $createResponse->getHeaderLine('content-type'));
+
+        $createResponseData = json_decode($createResponse->getBody()->getContents(), true);
+        self::assertIsArray($createResponseData);
+        self::assertArrayHasKey('id', $createResponseData);
+        self::assertTrue(Ulid::isValid($createResponseData['id']));
+        $jobId = $createResponseData['id'];
+
+        $getResponse = self::$staticApplicationClient->makeGetJobRequest($apiToken, $jobId);
+
+        self::assertSame(200, $getResponse->getStatusCode());
+        self::assertSame('application/json', $getResponse->getHeaderLine('content-type'));
+
+        $job = $jobRepository->find($jobId);
+        self::assertInstanceOf(Job::class, $job);
+
+        $responseData = json_decode($getResponse->getBody()->getContents(), true);
+
+        self::assertEquals($job->jsonSerialize(), $responseData);
+    }
+}
